+++ src/components/markdown-renderer/markdown-renderer.tsx
@@ -67,6 +67,7 @@ import { MermaidReplacer } from './replace-components/mermaid/mermaid-replacer'
 import { PdfReplacer } from './replace-components/pdf/pdf-replacer'
 import { PossibleWiderReplacer } from './replace-components/possible-wider/possible-wider-replacer'
 import { QuoteOptionsReplacer } from './replace-components/quote-options/quote-options-replacer'
+import { SequenceDiagramReplacer } from './replace-components/sequence-diagram/sequence-diagram-replacer'
 import { TaskListReplacer } from './replace-components/task-list/task-list-replacer'
 import { VimeoReplacer } from './replace-components/vimeo/vimeo-replacer'
 import { YoutubeReplacer } from './replace-components/youtube/youtube-replacer'
@@ -350,6 +351,7 @@ export const MarkdownRenderer: React.FC<MarkdownRendererProps> = ({
       new AbcReplacer(),
       new PdfReplacer(),
       new ImageReplacer(),
+      new SequenceDiagramReplacer(),
       new CsvReplacer(),
       new FlowchartReplacer(),
       new MermaidReplacer(),
@@ -380,7 +382,7 @@ export const MarkdownRenderer: React.FC<MarkdownRendererProps> = ({
             </Trans>
           </Alert>
         </ShowIf>
-        <div ref={documentElement} className={'markdown-body d-flex flex-column align-items-center'}>
+        <div ref={documentElement} className={'markdown-body w-100 d-flex flex-column align-items-center'}>
           {markdownReactDom}
         </div>
       </div>